Add http and epub modules
