+++ src/extractor.rs
@@ -1,10 +1,6 @@
 use std::collections::HashMap;
 
-use async_std::fs::File;
-use async_std::io::prelude::*;
-use async_std::task;
 use kuchiki::{traits::*, NodeRef};
-use url::Url;
 
 use crate::moz_readability::{MetaData, Readability};
 
@@ -51,8 +47,8 @@ impl Extractor {
     }
 
     /// Traverses the DOM tree of the content and retrieves the IMG URLs
-    fn extract_img_urls(&mut self) {
-        if let Some(content_ref) = &self.readability.article_node {
+    pub fn extract_img_urls(&mut self) {
+        if let Some(content_ref) = &self.article {
             for img_ref in content_ref.select("img").unwrap() {
                 img_ref.as_node().as_element().map(|img_elem| {
                     img_elem.attributes.borrow().get("src").map(|img_url| {
@@ -65,120 +61,12 @@ impl Extractor {
         }
     }
 
-    pub async fn download_images(&mut self, article_origin: &Url) -> async_std::io::Result<()> {
-        let mut async_download_tasks = Vec::with_capacity(self.img_urls.len());
-        self.extract_img_urls();
-        if self.img_urls.len() > 0 {
-            println!("Downloading images...");
-        }
-        for img_url in &self.img_urls {
-            let img_url = img_url.0.clone();
-            let abs_url = get_absolute_url(&img_url, article_origin);
-
-            async_download_tasks.push(task::spawn(async move {
-                let mut img_response = surf::Client::new()
-                    // The middleware has been temporarily commented out because it happens
-                    // to affect downloading images when there is no redirecting
-                    // .with(surf::middleware::Redirect::default())
-                    .get(&abs_url)
-                    .await
-                    .expect("Unable to retrieve file");
-                let img_content: Vec<u8> = img_response.body_bytes().await.unwrap();
-                let img_mime = img_response
-                    .content_type()
-                    .map(|mime| mime.essence().to_string());
-                let img_ext = img_response
-                    .content_type()
-                    .map(|mime| map_mime_subtype_to_ext(mime.subtype()).to_string())
-                    .unwrap();
-                let mut img_path = std::env::temp_dir();
-                img_path.push(format!("{}.{}", hash_url(&abs_url), &img_ext));
-                let mut img_file = File::create(&img_path)
-                    .await
-                    .expect("Unable to create file");
-                img_file
-                    .write_all(&img_content)
-                    .await
-                    .expect("Unable to save to file");
-
-                (
-                    img_url,
-                    img_path
-                        .file_name()
-                        .map(|os_str_name| {
-                            os_str_name
-                                .to_str()
-                                .expect("Unable to get image file name")
-                                .to_string()
-                        })
-                        .unwrap(),
-                    img_mime,
-                )
-            }));
-        }
-
-        self.img_urls.clear();
-
-        for async_task in async_download_tasks {
-            let (img_url, img_path, img_mime) = async_task.await;
-            // Update the image sources
-            let img_ref = self
-                .readability
-                .article_node
-                .as_mut()
-                .expect("Unable to get mutable ref")
-                .select_first(&format!("img[src='{}']", img_url))
-                .expect("Image node does not exist");
-            let mut img_node = img_ref.attributes.borrow_mut();
-            *img_node.get_mut("src").unwrap() = img_path.clone();
-            // srcset is removed because readers such as Foliate then fail to display
-            // the image already downloaded and stored in src
-            img_node.remove("srcset");
-            self.img_urls.push((img_path, img_mime));
-        }
-        Ok(())
-    }
-
     pub fn article(&self) -> Option<&NodeRef> {
         self.article.as_ref()
     }
 
     pub fn metadata(&self) -> &MetaData {
         &self.readability.metadata
-    }
-}
-
-/// Utility for hashing URLs. This is used to help store files locally with unique values
-fn hash_url(url: &str) -> String {
-    format!("{:x}", md5::compute(url.as_bytes()))
-}
-
-/// Handles getting the extension from a given MIME subtype.
-fn map_mime_subtype_to_ext(subtype: &str) -> &str {
-    if subtype == ("svg+xml") {
-        return "svg";
-    } else if subtype == "x-icon" {
-        "ico"
-    } else {
-        subtype
-    }
-}
-
-fn get_absolute_url(url: &str, request_url: &Url) -> String {
-    if Url::parse(url).is_ok() {
-        url.to_owned()
-    } else if url.starts_with("/") {
-        Url::parse(&format!(
-            "{}://{}",
-            request_url.scheme(),
-            request_url.host_str().unwrap()
-        ))
-        .unwrap()
-        .join(url)
-        .unwrap()
-        .into_string()
-    } else {
-        request_url.join(url).unwrap().into_string()
     }
 }
 
@@ -276,21 +164,6 @@ mod test {
         assert_eq!(
             vec![("http://example.com/img.jpg".to_string(), None)],
             extractor.img_urls
-        );
-    }
-
-    #[test]
-    fn test_map_mime_type_to_ext() {
-        let mime_subtypes = vec![
-            "apng", "bmp", "gif", "x-icon", "jpeg", "png", "svg+xml", "tiff", "webp",
-        ];
-        let exts = mime_subtypes
-            .into_iter()
-            .map(|mime_type| map_mime_subtype_to_ext(mime_type))
-            .collect::<Vec<_>>();
-        assert_eq!(
-            vec!["apng", "bmp", "gif", "ico", "jpeg", "png", "svg", "tiff", "webp"],
-            exts
         );
     }
 }
